Refactor/reorganize/rename trigger testing stuff - in preparation for invocation testing
