+++ app/models/behaviors/testing/TestEvent.scala
@@ -1,0 +1,5 @@
+package models.behaviors.testing
+
+import models.behaviors.events.MessageEvent
+
+case class TestEvent(context: TestMessageContext) extends MessageEvent
