+++ app/views/editBehavior.scala.html
@@ -12,7 +12,7 @@
     routes.javascript.BehaviorEditorController.save,
     routes.javascript.BehaviorEditorController.regexValidationErrorsFor,
     routes.javascript.BehaviorEditorController.versionInfoFor,
-    routes.javascript.BehaviorEditorController.test,
+    routes.javascript.BehaviorEditorController.testTriggers,
     routes.javascript.EnvironmentVariablesController.submit,
     routes.javascript.OAuth2ApplicationController.newApp
   )
